feat: Backend Spring Boot local funcional com Cloud SQL

+++ src/main/java/br/com/atletahub/atletahub_backend/dto/perfil/DadosAtualizacaoPerfilAtleta.java
@@ -2,14 +2,15 @@ package br.com.atletahub.atletahub_backend.dto.perfil;
 
 import jakarta.validation.constraints.Pattern;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 public record DadosAtualizacaoPerfilAtleta(
         String nome,
         String email,
         Integer idade,
-        Double altura,
-        Double peso,
+        BigDecimal altura,
+        BigDecimal peso,
         String modalidade,
         String competicoesTitulos,
         String redesSocial,
